requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bce-python-sdk==0.9.55
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.3.1
colorlog==6.10.1
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
Django==5.2.3
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
easyocr==1.7.2
filelock==3.18.0
fsspec==2025.5.1
future==1.0.0
h11==0.16.0
h2==4.2.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
imagesize==1.4.1
incremental==24.7.2
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
modelscope==1.33.0
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==25.0
paddlex==3.3.11
pandas==2.3.3
pillow==11.3.0
prettytable==3.17.0
priority==1.3.0
protobuf==6.33.2
psutil==7.1.3
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.9.0
pyOpenSSL==25.1.0
pypdfium2==5.2.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.16.0
service-identity==24.2.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
Twisted==25.5.0
txaio==25.6.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14
zope.interface==7.2
